+++ app/src/features/sessionBook/index.ts
@@ -1,2 +1,3 @@
 export { sessionRoutes } from "./routes";
 export * from "./types";
+export { SessionsOnboardingView } from "./screens/SessionsListScreen/components/OnboardingView/SessionsOnboardingView";
