+++ app/src/features/sessionBook/screens/SessionsListScreen/components/OnboardingView/sessionsOnboardingView.scss
@@ -6,8 +6,7 @@
 
 .onboarding-content-container {
   .settings-row {
-    margin-right: 9px;
-    margin-top: 9px;
+    padding: 1rem 1rem 0 0;
   }
 
   .settings-btn {
