+++ app/src/routes/desktopSessionRoutes.tsx
@@ -1,8 +1,8 @@
 import PATHS from "config/constants/sub/paths";
+import { SessionListScreenContainer } from "features/sessionBook/screens/SessionsListScreen/SessionListScreenContainer";
 import { Navigate, RouteObject } from "react-router-dom";
 import { DesktopSessionsContainer } from "views/containers/DesktopSessionsContainer";
 import { DraftSessionViewer, SavedSessionViewer } from "views/features/sessions/SessionViewer";
-import SessionsIndexPageContainer from "views/features/sessions/SessionsIndexPageContainer";
 import NetworkSessionsIndexPage from "views/features/sessions/SessionsIndexPageContainer/NetworkSessions";
 import NetworkSessionViewer from "views/features/sessions/SessionsIndexPageContainer/NetworkSessions/NetworkSessionViewer";
 
@@ -26,7 +26,7 @@ export const desktopSessionsRoutes: RouteObject[] = [
       },
       {
         path: PATHS.SESSIONS.DESKTOP.WEB_SESSIONS_WRAPPER.RELATIVE,
-        element: <SessionsIndexPageContainer />,
+        element: <SessionListScreenContainer />,
       },
       {
         path: PATHS.SESSIONS.DESKTOP.SAVED_WEB_SESSION_VIEWER.ABSOLUTE + "/:id",
