+++ app/src/views/features/sessions/SessionsIndexPageContainer/NetworkSessions/NetworkSessionViewer/NetworkSessionsOnboardingView.tsx
@@ -1,0 +1,52 @@
+import { Divider } from "antd";
+import React from "react";
+import { ImportHarModalButton } from "../ImportHarModalButton";
+import { CheckOutlined } from "@ant-design/icons";
+import { Typography } from "antd";
+
+const CheckItem: React.FC<{ label: string }> = ({ label }) => {
+  return (
+    <div>
+      <CheckOutlined style={{ marginRight: "8px", fontSize: "16px", color: "#228B22" }} />
+      <span>{label}</span>
+    </div>
+  );
+};
+
+export const NewtorkSessionsOnboardingView: React.FC<{}> = () => {
+  const { Title, Text } = Typography;
+  return (
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "column",
+        textAlign: "center",
+        height: "100%",
+        margin: "30px",
+      }}
+    >
+      <Title level={1}>Record &amp; Replay your browsing sessions</Title>
+      <Text type="secondary">
+        <div>Record your network sessions and Share with others for offline review or debugging.</div>
+      </Text>
+      <div>
+        {/* <HarImportModal onSaved={stableNavigate} /> */}
+        <ImportHarModalButton />
+      </div>
+      <Divider />
+      <Text type="secondary">
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "space-evenly",
+            fontWeight: "bold",
+          }}
+        >
+          <CheckItem label="Faster Debugging" />
+          <CheckItem label="No need to reproduce" />
+          <CheckItem label="Strict Privacy" />
+        </div>
+      </Text>
+    </div>
+  );
+};
