+++ app/src/views/features/sessions/SessionsIndexPageContainer/NetworkSessions/index.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from "react";
 import NetworkSessionsList from "./NetworkSessionsList";
 import { NetworkSessionRecord } from "./types";
-import OnboardingView, { OnboardingTypes } from "../SessionsIndexPage/OnboardingView";
+
 import { submitAttrUtil } from "utils/AnalyticsUtils";
 import APP_CONSTANTS from "config/constants";
+import { NewtorkSessionsOnboardingView } from "./NetworkSessionViewer/NetworkSessionsOnboardingView";
 const TRACKING = APP_CONSTANTS.GA_EVENTS;
 
 const NetworkSessionsIndexPage: React.FC<{}> = () => {
@@ -33,7 +34,7 @@ const NetworkSessionsIndexPage: React.FC<{}> = () => {
   return networkSessions.length ? (
     <NetworkSessionsList networkSessionsMetadata={networkSessions} />
   ) : (
-    <OnboardingView type={OnboardingTypes.NETWORK} />
+    <NewtorkSessionsOnboardingView />
   );
 };
