+++ app/src/views/features/sessions/index.ts
@@ -1,5 +1,3 @@
 // import lazyload from "utils/lazyload";
 
 // export default lazyload(() => import(/* webpackChunkName: "SessionsHomeView" */ "./SessionsHomeView"));
-
-export { default } from "./SessionsHomeView";
